+++ src/candlewick/core/Device.cpp
@@ -1,7 +1,7 @@
 #include "Device.h"
 #include "errors.h"
 
-#include <SDL3/SDL_log.h>
+#include <spdlog/spdlog.h>
 
 namespace candlewick {
 
@@ -25,8 +25,7 @@ void Device::create(SDL_GPUShaderFormat format_flags, bool debug_mode) {
   _device = SDL_CreateGPUDevice(format_flags, debug_mode, nullptr);
   if (!_device)
     throw RAIIException(SDL_GetError());
-  const char *driver = SDL_GetGPUDeviceDriver(_device);
-  SDL_Log("Device driver: %s", driver);
+  spdlog::info("Device driver: {:s}", driverName());
 }
 
 const char *Device::driverName() const noexcept {
@@ -34,8 +33,10 @@ const char *Device::driverName() const noexcept {
 }
 
 void Device::destroy() noexcept {
-  if (_device)
+  if (_device) {
+    spdlog::debug("Destroying GPU device");
     SDL_DestroyGPUDevice(_device);
+  }
   _device = nullptr;
 }