+++ src/candlewick/core/Renderer.h
@@ -1,5 +1,0 @@
-#include "candlewick/deprecated.h"
-#include "RenderContext.h"
-
-CANDLEWICK_DEPRECATED_HEADER(
-    "Include \'<candlewick/RenderContext.h>\' instead.")